Avoid retain cycle so we don't leak memory. Delegates should almost always be weak.
In this case ViewPost and EditPost owns a TripData object, i.e retains it. But they are also postDel on that object. And since that is a strong reference TripData retains ViewPost and EditPost meaning that we have a retain cycle and memory will never be released.

In the same way MyTrips and PostsCollectionViewController owns a TripData object. But they are also dataDel on that same object, which if being a strong refernce creates a retain cycle.